requirements.txt
Django==5.2.3
python-telegram-bot==22.1
python-dotenv==1.1.1
django-dotenv==1.4.2